Check for the existance of the table before making queries against it. It might really have been deleted.

+++ docs/upgrades/upgrade_7.7.6-7.7.7.pl
@@ -52,14 +52,17 @@ finish($session); # this line required
 sub removeDanglingOldRssAssets {
     my $session = shift;
     print "\tChecking for uses of RSSCapable...\n" unless $quiet;
-    my @rssCapableClasses = $session->db->buildArray('SELECT className FROM RSSCapable INNER JOIN asset ON RSSCapable.assetId=asset.assetId GROUP BY className');
-    if (@rssCapableClasses) {
-        warn "\t\tThis site is using the assets\n\t\t\t" . join(', ', @rssCapableClasses) . "\n\t\twhich use the RSSCapable class!  Support RSSCapable has been dropped and it will no longer be maintained.\n";
-    }
-    else {
-        print "\t\tNot used, removing leftover assets, if any.\n" unless $quiet;
-        $session->db->write(q|DELETE FROM assetData WHERE assetId IN (SELECT assetId FROM ASSET WHERE className="WebGUI::Asset::RssFromParent")|);
-        $session->db->write(q|DELETE FROM asset WHERE className = "WebGUI::Asset::RssFromParent"|);
+    my $peek = $session->db->dbh->table_info(undef, undef, 'RSSCapable');
+    if ($peek->fetchrow_hashref()) {
+        my @rssCapableClasses = $session->db->buildArray('SELECT className FROM RSSCapable INNER JOIN asset ON RSSCapable.assetId=asset.assetId GROUP BY className');
+        if (@rssCapableClasses) {
+            warn "\t\tThis site is using the assets\n\t\t\t" . join(', ', @rssCapableClasses) . "\n\t\twhich use the RSSCapable class!  Support RSSCapable has been dropped and it will no longer be maintained.\n";
+        }
+        else {
+            print "\t\tNot used, removing leftover assets, if any.\n" unless $quiet;
+            $session->db->write(q|DELETE FROM assetData WHERE assetId IN (SELECT assetId FROM ASSET WHERE className="WebGUI::Asset::RssFromParent")|);
+            $session->db->write(q|DELETE FROM asset WHERE className = "WebGUI::Asset::RssFromParent"|);
+        }
     }
     print "\tDone.\n" unless $quiet;
 }
